ai_services/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
# PyMuPDF
google-cloud-vision
python-dotenv
requests
aiofiles
python-dotenv
paddleocr>=2.0.1
PyMuPDF==1.21.1